feat: Update Rating column type from byte to int in Record table

+++ Infrastructure/Migrations/20250424190441_AddRating.cs
@@ -41,7 +41,7 @@ namespace Infrastructure.Migrations
                 oldMaxLength: 150,
                 oldNullable: true);
 
-            migrationBuilder.AddColumn<byte>(
+            migrationBuilder.AddColumn<int>(
                 name: "Rating",
                 table: "Record",
                 type: "integer",
